\name{cropsim-package}
\alias{cropsim-package}
\alias{cropsim}
\docType{package}
\title{
What the package does (short line)
	Dynamic crop simulation models (disease models have been moved to companion package simdisease)
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab simplant\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-11-8\cr
License: \tab GPL v3\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Robert J. Hijmans

Maintainer: <r.hijmans@gmail.com>

}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }

\seealso{
\code{\link[simdisease:simdisease-package]{simdisease}} 
}
\examples{

}
